articular-es: entities as components + migrated ID to components
