+++ articular-es/src/main/java/articular/core/Entity.java
@@ -31,8 +31,8 @@
 
 package articular.core;
 
+import articular.core.component.Component;
 import articular.core.system.manager.EntityComponentManager;
-import articular.util.Validatable;
 
 /**
  * Offers a static ID provider for the {@link MemoryMap.EntityComponentMap}
@@ -42,7 +42,7 @@ import articular.util.Validatable;
  * @see MemoryMap.EntityComponentMap
  * @see EntityComponentManager
  */
-public final class Entity implements Validatable {
+public final class Entity implements Component {
 
     private final String name;
 
@@ -50,52 +50,8 @@ public final class Entity implements Validatable {
         this.name = name;
     }
 
+    @Override
     public Id getId() {
         return new Entity.Id((name.hashCode() >>> 16) ^ name.hashCode());
-    }
-
-    /**
-     * Represents the component identifier of type number, this encapsulates
-     * a constant long value to accommodate larger number of game entity components.
-     */
-    public static final class Id extends Number {
-        private final int id;
-
-        /**
-         * Instantiates a game entity component identifier object.
-         *
-         * @param id the identifier in long format
-         */
-        public Id(final int id) {
-            this.id = id;
-        }
-
-        /**
-         * Retrieves the identifier in a long format.
-         *
-         * @return the component identifier in longs
-         */
-        @Override
-        public int intValue() {
-            return id;
-        }
-
-        @Deprecated
-        @Override
-        public long longValue() {
-            throw new UnsupportedOperationException("Deprecated call, use \"Component.Id#intValue()\"");
-        }
-
-        @Deprecated
-        @Override
-        public float floatValue() {
-            throw new UnsupportedOperationException("Deprecated call, use \"Component.Id#intValue()\"");
-        }
-
-        @Deprecated
-        @Override
-        public double doubleValue() {
-            throw new UnsupportedOperationException("Deprecated call, use \"Component.Id#intValue()\"");
-        }
     }
 }
